sim: split the simulator in smaller files
