+++ tests/rpl.rs
@@ -26,15 +26,15 @@ fn root_node_only(#[case] mop: RplModeOfOperation) {
 
     sim.run(Duration::from_millis(500), ONE_HOUR);
 
-    assert!(!sim.messages.is_empty());
+    assert!(!sim.msgs().is_empty());
 
     // In 1 hour, a root node will transmit around 10 messages.
-    let dio_count = sim.messages.iter().filter(|m| m.is_dio().unwrap()).count();
+    let dio_count = sim.msgs().iter().filter(|m| m.is_dio()).count();
     assert!(dio_count == 9 || dio_count == 10 || dio_count == 11);
 
     // There should only be DIO's.
-    for msg in &sim.messages {
-        assert!(msg.is_dio().unwrap());
+    for msg in sim.msgs() {
+        assert!(msg.is_dio());
     }
 }
 
@@ -51,15 +51,15 @@ fn normal_node_without_dodag(#[case] mop: RplModeOfOperation) {
 
     sim.run(Duration::from_millis(500), ONE_HOUR);
 
-    assert!(!sim.messages.is_empty());
+    assert!(!sim.msgs().is_empty());
 
     // In 1 hour, around 60 DIS messages are transmitted by 1 node.
-    let dis_count = sim.messages.iter().filter(|m| m.is_dis().unwrap()).count();
+    let dis_count = sim.msgs().iter().filter(|m| m.is_dis()).count();
     assert!(dis_count == 59 || dis_count == 60 || dis_count == 61);
 
     // There should only be DIS messages.
-    for msg in &sim.messages {
-        assert!(msg.is_dis().unwrap());
+    for msg in sim.msgs() {
+        assert!(msg.is_dis());
     }
 }
 
@@ -78,9 +78,9 @@ fn root_and_normal_node(#[case] mop: RplModeOfOperation) {
 
     sim.run(Duration::from_millis(500), Duration::from_secs(60 * 15));
 
-    assert!(!sim.messages.is_empty());
+    assert!(!sim.msgs().is_empty());
 
-    let dio_count = sim.messages.iter().filter(|m| m.is_dio().unwrap()).count();
+    let dio_count = sim.msgs().iter().filter(|m| m.is_dio()).count();
 
     assert!(dio_count > 12 && dio_count < 17);
 
@@ -88,12 +88,8 @@ fn root_and_normal_node(#[case] mop: RplModeOfOperation) {
         RplModeOfOperation::NonStoringMode
         | RplModeOfOperation::StoringMode
         | RplModeOfOperation::StoringModeWithMulticast => {
-            let dao_count = sim.messages.iter().filter(|m| m.is_dao().unwrap()).count();
-            let dao_ack_count = sim
-                .messages
-                .iter()
-                .filter(|m| m.is_dao_ack().unwrap())
-                .count();
+            let dao_count = sim.msgs().iter().filter(|m| m.is_dao()).count();
+            let dao_ack_count = sim.msgs().iter().filter(|m| m.is_dao_ack()).count();
 
             assert_eq!(dao_count, 1);
             assert_eq!(dao_ack_count, 1);
@@ -101,17 +97,17 @@ fn root_and_normal_node(#[case] mop: RplModeOfOperation) {
         _ => (),
     }
 
-    for msg in &sim.messages {
+    for msg in sim.msgs() {
         match mop {
             // In MOP0, all messages should be DIOs. A node only transmits its first DIS after 5
             // seconds. The first DIO from the root is transmitted after 2 - 4 seconds after the
             // start. Therefore, there should never be a DIS in the messages.
-            RplModeOfOperation::NoDownwardRoutesMaintained => assert!(msg.is_dio().unwrap()),
+            RplModeOfOperation::NoDownwardRoutesMaintained => assert!(msg.is_dio()),
             // In MOP1, MOP2, MOP3, DAOs and DAO-ACKs are also transmitted.
             RplModeOfOperation::NonStoringMode
             | RplModeOfOperation::StoringMode
             | RplModeOfOperation::StoringModeWithMulticast => {
-                assert!(msg.is_dio().unwrap() || msg.is_dao().unwrap() || msg.is_dao_ack().unwrap())
+                assert!(msg.is_dio() || msg.is_dao() || msg.is_dao_ack())
             }
         }
     }
@@ -126,34 +122,34 @@ fn root_and_normal_node_moved_out_of_range(#[case] mop: RplModeOfOperation) {
 
     sim.run(Duration::from_millis(100), ONE_HOUR);
 
-    assert!(!sim.messages.is_empty());
+    assert!(!sim.msgs().is_empty());
 
     // We check that a node is connect to the DODAG, meaning there should be no DIS messages.
-    for msg in &sim.messages {
+    for msg in sim.msgs() {
         match mop {
             // In MOP0, all messages should be DIOs. A node only transmits its first DIS after 5
             // seconds. The first DIO from the root is transmitted after 2 - 4 seconds after the
             // start. Therefore, there should never be a DIS in the messages.
-            RplModeOfOperation::NoDownwardRoutesMaintained => assert!(msg.is_dio().unwrap()),
+            RplModeOfOperation::NoDownwardRoutesMaintained => assert!(msg.is_dio()),
             // In MOP1, MOP2, MOP3, DAOs and DAO-ACKs are also transmitted.
             RplModeOfOperation::NonStoringMode
             | RplModeOfOperation::StoringMode
             | RplModeOfOperation::StoringModeWithMulticast => {
-                assert!(msg.is_dio().unwrap() || msg.is_dao().unwrap() || msg.is_dao_ack().unwrap())
+                assert!(msg.is_dio() || msg.is_dao() || msg.is_dao_ack())
             }
         }
     }
 
-    sim.messages.clear();
+    sim.clear_msgs();
 
     // Move the node far from the root node.
-    sim.nodes[1].set_position(sim::Position((1000., 0.)));
+    sim.nodes_mut()[1].set_position(sim::Position((1000., 0.)));
 
     sim.run(Duration::from_millis(400), ONE_HOUR);
 
     match mop {
         RplModeOfOperation::NonStoringMode | RplModeOfOperation::StoringMode => {
-            let dao_count = sim.messages.iter().filter(|m| m.is_dao().unwrap()).count();
+            let dao_count = sim.msgs().iter().filter(|m| m.is_dao()).count();
             assert!(dao_count < 5);
         }
         _ => {}
@@ -161,11 +157,11 @@ fn root_and_normal_node_moved_out_of_range(#[case] mop: RplModeOfOperation) {
 
     // When a node leaves a DODAG, it multicasts an INFINITE rank DIO.
     let infinite_rank_dio_count = sim
-        .messages
+        .msgs()
         .iter()
         .filter(|m| {
-            if m.is_dio().unwrap() {
-                let icmp = m.icmp().unwrap().unwrap();
+            if m.is_dio() {
+                let icmp = m.icmp().unwrap();
                 let Icmpv6Repr::Rpl(RplRepr::DodagInformationObject(dio)) = icmp else {
                     return false;
                 };
@@ -178,49 +174,44 @@ fn root_and_normal_node_moved_out_of_range(#[case] mop: RplModeOfOperation) {
 
     assert!(infinite_rank_dio_count == 1);
 
-    for msg in &sim.messages {
+    for msg in sim.msgs() {
         match mop {
             // There should be no DAO or DAO-ACK, however, it should containt DIS's.
             RplModeOfOperation::NoDownwardRoutesMaintained => {
-                assert!(msg.is_dio().unwrap() || msg.is_dis().unwrap())
+                assert!(msg.is_dio() || msg.is_dis())
             }
             RplModeOfOperation::NonStoringMode
             | RplModeOfOperation::StoringMode
             | RplModeOfOperation::StoringModeWithMulticast => {
-                assert!(msg.is_dio().unwrap() || msg.is_dis().unwrap() || msg.is_dao().unwrap());
+                assert!(msg.is_dio() || msg.is_dis() || msg.is_dao());
             }
         }
     }
 
-    sim.messages.clear();
+    sim.clear_msgs();
 
     // Move the node back in range of the root node.
-    sim.nodes[1].set_position(sim::Position((100., 0.)));
+    sim.nodes_mut()[1].set_position(sim::Position((100., 0.)));
 
     sim.run(Duration::from_millis(100), ONE_HOUR);
 
     // NOTE: in rare cases, I don't know why, 2 DIS messages are transmitted instead of just 1.
-    let dis_count = sim.messages.iter().filter(|m| m.is_dis().unwrap()).count();
+    let dis_count = sim.msgs().iter().filter(|m| m.is_dis()).count();
     assert!(dis_count < 3);
 
-    for msg in &sim.messages {
+    for msg in sim.msgs() {
         match mop {
             // In MOP0, all messages should be DIOs. A node only transmits its first DIS after 5
             // seconds. The first DIO from the root is transmitted after 2 - 4 seconds after the
             // start. Therefore, there should never be a DIS in the messages.
             RplModeOfOperation::NoDownwardRoutesMaintained => {
-                assert!(msg.is_dio().unwrap() || msg.is_dis().unwrap())
+                assert!(msg.is_dio() || msg.is_dis())
             }
             // In MOP1, MOP2, MOP3, DAOs and DAO-ACKs are also transmitted.
             RplModeOfOperation::NonStoringMode
             | RplModeOfOperation::StoringMode
             | RplModeOfOperation::StoringModeWithMulticast => {
-                assert!(
-                    msg.is_dis().unwrap()
-                        || msg.is_dio().unwrap()
-                        || msg.is_dao().unwrap()
-                        || msg.is_dao_ack().unwrap()
-                )
+                assert!(msg.is_dis() || msg.is_dio() || msg.is_dao() || msg.is_dao_ack())
             }
         }
     }
@@ -233,39 +224,39 @@ fn root_and_normal_node_moved_out_of_range(#[case] mop: RplModeOfOperation) {
 fn message_forwarding_to_root(#[case] mop: RplModeOfOperation) {
     let mut sim = sim::topology(sim::NetworkSim::new(), mop, 1, 2);
 
-    let dst_addr = sim.nodes[0].ip_address;
-    sim::udp_receiver_node(&mut sim.nodes[0], 1234);
-    sim::udp_sender_node(&mut sim.nodes[2], 1234, dst_addr);
+    let dst_addr = sim.nodes()[0].ip_address;
+    sim::udp_receiver_node(&mut sim.nodes_mut()[0], 1234);
+    sim::udp_sender_node(&mut sim.nodes_mut()[2], 1234, dst_addr);
 
     sim.init();
     sim.run(Duration::from_millis(500), ONE_HOUR);
 
-    assert!(!sim.messages.is_empty());
+    assert!(!sim.msgs().is_empty());
 
-    let dio_count = sim.messages.iter().filter(|m| m.is_dio().unwrap()).count();
+    let dio_count = sim.msgs().iter().filter(|m| m.is_dio()).count();
     assert!(dio_count > 27 && dio_count < 33);
 
     // We transmit a message every 60 seconds. We simulate for 1 hour, so the node will transmit
     // 59 messages. The node is not in range of the destination (which is the root). There is one
     // node inbetween that has to forward it. Thus, it is forwarding 59 messages.
-    let udp_count = sim.messages.iter().filter(|m| m.is_udp().unwrap()).count();
+    let udp_count = sim.msgs().iter().filter(|m| m.is_udp()).count();
     assert!(udp_count >= 59 * 2 && udp_count <= 60 * 2);
 
-    for msg in &sim.messages {
+    for msg in sim.msgs() {
         match mop {
             RplModeOfOperation::NoDownwardRoutesMaintained => {
-                assert!(msg.is_dis().unwrap() || msg.is_dio().unwrap() || msg.is_udp().unwrap())
+                assert!(msg.is_dis() || msg.is_dio() || msg.is_udp())
             }
             // In MOP1, MOP2, MOP3, DAOs and DAO-ACKs are also transmitted.
             RplModeOfOperation::NonStoringMode
             | RplModeOfOperation::StoringMode
             | RplModeOfOperation::StoringModeWithMulticast => {
                 assert!(
-                    msg.is_dis().unwrap()
-                        || msg.is_dio().unwrap()
-                        || msg.is_dao().unwrap()
-                        || msg.is_dao_ack().unwrap()
-                        || msg.is_udp().unwrap()
+                    msg.is_dis()
+                        || msg.is_dio()
+                        || msg.is_dao()
+                        || msg.is_dao_ack()
+                        || msg.is_udp()
                 )
             }
         }
@@ -279,22 +270,22 @@ fn message_forwarding_to_root(#[case] mop: RplModeOfOperation) {
 fn message_forwarding_up_and_down(#[case] mop: RplModeOfOperation) {
     let mut sim = sim::topology(sim::NetworkSim::new(), mop, 2, 2);
 
-    let dst_addr = sim.nodes[3].ip_address;
-    sim::udp_receiver_node(&mut sim.nodes[3], 1234);
-    sim::udp_sender_node(&mut sim.nodes[4], 1234, dst_addr);
+    let dst_addr = sim.nodes()[3].ip_address;
+    sim::udp_receiver_node(&mut sim.nodes_mut()[3], 1234);
+    sim::udp_sender_node(&mut sim.nodes_mut()[4], 1234, dst_addr);
 
     sim.init();
     sim.run(Duration::from_millis(500), Duration::from_secs(60 * 15));
 
-    assert!(!sim.messages.is_empty());
+    assert!(!sim.msgs().is_empty());
 
-    let dio_count = sim.messages.iter().filter(|m| m.is_dio().unwrap()).count();
+    let dio_count = sim.msgs().iter().filter(|m| m.is_dio()).count();
     assert!(dio_count >= 30 && dio_count <= 40);
 
     // We transmit a message every 60 seconds. We simulate for 1 hour, so the node will transmit
     // 59 messages. The node is not in range of the destination (which is the root). There is one
     // node inbetween that has to forward it. Thus, it is forwarding 59 messages.
-    let udp_count = sim.messages.iter().filter(|m| m.is_udp().unwrap()).count();
+    let udp_count = sim.msgs().iter().filter(|m| m.is_udp()).count();
     match mop {
         RplModeOfOperation::NoDownwardRoutesMaintained => {
             assert!(udp_count >= 14 * 2 && udp_count <= 15 * 2);
@@ -306,41 +297,41 @@ fn message_forwarding_up_and_down(#[case] mop: RplModeOfOperation) {
         }
     }
 
-    for msg in &sim.messages {
+    for msg in sim.msgs() {
         match mop {
             RplModeOfOperation::NoDownwardRoutesMaintained => {
-                assert!(msg.is_dis().unwrap() || msg.is_dio().unwrap() || msg.is_udp().unwrap())
+                assert!(msg.is_dis() || msg.is_dio() || msg.is_udp())
             }
             // In MOP1, MOP2, MOP3, DAOs and DAO-ACKs are also transmitted.
             RplModeOfOperation::NonStoringMode
             | RplModeOfOperation::StoringMode
             | RplModeOfOperation::StoringModeWithMulticast => {
                 assert!(
-                    msg.is_dis().unwrap()
-                        || msg.is_dio().unwrap()
-                        || msg.is_dao().unwrap()
-                        || msg.is_dao_ack().unwrap()
-                        || msg.is_udp().unwrap()
+                    msg.is_dis()
+                        || msg.is_dio()
+                        || msg.is_dao()
+                        || msg.is_dao_ack()
+                        || msg.is_udp()
                 )
             }
         }
     }
 
     // All UDP, DAO, DAO-ACK packets should have a HBH or a source routing header
-    sim.messages
+    sim.msgs()
         .iter()
-        .filter(|m| m.is_udp().unwrap() || m.is_dao().unwrap() || m.is_dao_ack().unwrap())
-        .for_each(|m| assert!(m.has_hbh().unwrap() || m.has_routing().unwrap()));
+        .filter(|m| m.is_udp() || m.is_dao() || m.is_dao_ack())
+        .for_each(|m| assert!(m.has_hbh() || m.has_routing()));
 
     let dao_ack_packets_with_routing = sim
-        .messages
+        .msgs()
         .iter()
-        .filter(|m| m.is_dao_ack().unwrap() && m.has_routing().unwrap())
+        .filter(|m| m.is_dao_ack() && m.has_routing())
         .count();
     let dao_ack_packets_without_routing = sim
-        .messages
+        .msgs()
         .iter()
-        .filter(|m| m.is_dao_ack().unwrap() && !m.has_routing().unwrap())
+        .filter(|m| m.is_dao_ack() && !m.has_routing())
         .count();
 
     match mop {
